smart-todo-backend/requirements.txt
asgiref==3.9.0
dj-database-url==3.0.1
Django==5.2.4
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
